+++ NEWS.md
@@ -5,6 +5,7 @@ Improvements
 
 Bug Fixes
 * Fix issue where pdf render fails if headers or footers start with brackets, e.g. "[1]" ([#85](https://github.com/GSK-Biostatistics/docorator/issues/85))
+* Fix issue where chunk ids could be duplicated (#88)
 
 # docorator 0.5.2